Regras de negócio e outras funções

+++ GerTarefas.Application/Projects/Commands/DeleteProjectCommand.cs
@@ -19,10 +19,17 @@ public class DeleteTaskCommand : IRequest<Project>
         public async Task<Project> Handle(DeleteTaskCommand request,
                      CancellationToken cancellationToken)
         {
+            var tasks = await _unitOfWork.TaskProjectRepository.GetTasksByProjectId(request.Id);
+
+            var taskExist = tasks.Any(c => c.Status == Domain.Enum.StatusEnum.PENDENTE);
+            if (taskExist)
+            {
+                throw new InvalidOperationException("Regra:Não possivel remover projeto, você deve concluir ou remover as tarefas pendentes primeiro.");
+            }
             var deletedProject = await _unitOfWork.ProjectRepository.DeleteProject(request.Id);
 
             if (deletedProject is null)
-                throw new InvalidOperationException("Project not found");
+                throw new KeyNotFoundException("Projeto não encontrado");
 
             await _unitOfWork.CommitAsync();
             return deletedProject;
